Troubleshooting MOS
To make sure Mongoose OS  has the ability to create credentials and a 'Thing' for your account.  The AWS CLI Sigv4 configuration must have user credentials. - "Access key" and "Secret access key" with permissions to AWS IoT Core.  This user can be created in the IAM console and then configured with the AWS CLI